requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
pandas==2.2.2
openpyxl==3.1.2
pydantic==2.7.1
python-dotenv==1.0.1
tqdm==4.66.4
torch==2.3.0
transformers==4.41.2
